requirements.txt
django==3.1.4
djangorestframework==3.12.2
django-filter==2.4.0
django-bootstrap-v5